pyproject.toml
[project]
name = "foak-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.112.1",
    "logfire>=4.3.5",
    "pydantic-ai-slim[tavily]>=0.7.4",
    "pydantic-ai[tavily]>=0.7.4",
    "python-dotenv>=1.1.1",
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.10",
]
